move api tests to the api module
